add support to load puzzles from chess.com API
This adds puzzle mechanics with methods to check correct answer,
perform the next move and keep the puzzle status, if an attempt failed
the puzzle is marked as failed. It also includes UI change to visualize
whether the puzzle was solved and if player was successful or not. Also
include a new shortcut to clean user's input field.